+++ src/features/presence/Presence.scss
@@ -117,19 +117,6 @@
     margin-top: 20px;
 }
 
-.presence-avatar-tooltip-container {
-    text-align: center;
-}
-
-.presence-avatar-tooltip-name {
-    font-weight: bold;
-    font-size: 13px;
-}
-
-.presence-avatar-tooltip-event {
-    font-size: 11px;
-}
-
 .dropshadow-list,
 .dropshadow-list-with-buttons {
     overflow-x: hidden;
