+++ src/features/presence/PresenceAvatarTooltipContent.js.flow
@@ -1,5 +1,5 @@
 /**
- * Flowtype definitions for PresenceAvatarTooltip.tsx
+ * Flowtype definitions for PresenceAvatarTooltipContent.tsx
  * Generated by Flowgen from a Typescript Definition
  * Flowgen v1.10.0
  */
@@ -12,6 +12,7 @@ import {
 } from "react-intl";
 import messages from "./messages";
 import { determineInteractionMessage } from "./utils/presenceUtils";
+import "./PresenceAvatarTooltipContent.scss";
 export type Props = {
   name: string,
   interactedAt: number,
@@ -19,6 +20,6 @@ export type Props = {
   isActive?: boolean,
   ...
 } & WrappedComponentProps;
-declare var PresenceAvatarTooltip: any; // /* NO PRINT IMPLEMENTED: ArrowFunction */ any
-declare export { PresenceAvatarTooltip as PresenceAvatarTooltipComponent };
+declare function PresenceAvatarTooltipContent(x: Props): React$Node;
+declare export { PresenceAvatarTooltipContent as PresenceAvatarTooltipContentComponent };
 declare export default typeof /* NO PRINT IMPLEMENTED: CallExpression */ any;
